api-service/requirements.txt
aioredis==2.0.1
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
async-timeout==5.0.1
asyncpg==0.30.0
attrs==24.3.0
boto3==1.38.32
botocore==1.38.32
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
dateparser==1.2.0
decorator==5.1.1
Deprecated==1.2.18
dnspython==2.7.0
dotenv==0.9.9
duckdb==1.3.0
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.11
fastapi-cache2==0.2.2
fsspec==2024.12.0
GeoAlchemy2==0.17.1
geojson-pydantic==1.2.0
geopandas==1.0.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jwt==1.3.1
limits==5.2.0
logger==1.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
nest-asyncio==1.6.0
numpy==2.2.2
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pendulum==3.1.0
pika==1.3.2
platformdirs==4.3.6
psutil==6.1.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.4
pydantic-extra-types==2.10.3
pydantic-settings==2.8.1
pydantic_core==2.27.2
pygeoif==1.5.1
Pygments==2.19.1
PyJWT==2.10.1
pyogrio==0.10.0
pyproj==3.7.0
pystac==1.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
redis==6.2.0
referencing==0.35.1
regex==2024.11.6
rich==13.9.4
rich-toolkit==0.13.2
rpds-py==0.22.3
rsa==4.9
s3transfer==0.13.0
shapely==2.0.6
shellingham==1.5.4
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.39
SQLAlchemy-Utils==0.41.2
stac-pydantic==3.1.3
starlette==0.46.0
sympy==1.13.1
tenacity==8.2.2
typer==0.15.2
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3.1
ujson==5.10.0
urllib3==2.4.0
utils==1.0.2
uvicorn==0.34.0
watchfiles==1.0.4
wcwidth==0.2.13
websockets==15.0
wrapt==1.17.2
zipp==3.21.0